+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/container/mp3/Mp3ContainerProbe.java
@@ -64,7 +64,8 @@ public class Mp3ContainerProbe implements MediaContainerProbe {
           defaultOnNull(file.getIdv3Tag(ARTIST_TAG), UNKNOWN_ARTIST),
           file.getDuration(),
           reference.identifier,
-          !file.isSeekable()
+          !file.isSeekable(),
+          reference.identifier
       ));
     } finally {
       file.close();