+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/container/ogg/OggContainerProbe.java
@@ -46,7 +46,8 @@ public class OggContainerProbe implements MediaContainerProbe {
         UNKNOWN_ARTIST,
         Long.MAX_VALUE,
         reference.identifier,
-        true
+        true,
+        reference.identifier
     ));
   }